Fix Notification specs
Have had to refactor the comments factory and add the expected text to
the spec (these were half-finished at the time and contained the wrong
text). Have also decided to rename the method used
`Notification.new_comment_admin_notification` as most probably a new
notification will have to be written to notify users of replies to their
comments.