Add React dependencies
Rails 5 supports the use of React via Webpack (Webpacker gem) out of the
box so have decided to use this instead of the various React-Rails gems.

Have commented out a line that would have allowed for the entire
node_modules folder to be included in the app at a given time as the EJS
library constantly broke the app when `require_tree .` was used.

Have also made a change to the ClientsController#index action so as to
ensure that all clients can be rendered as a JSON list.