Fix issues with the modal and improve styling
Attributes for individual clients are now being passed to the modal as
part of the state meaning that the modal now displays the correct
information and closes correctly. Have also updated the styling for the
modal and client list so as to more closely resemble the old fancybox
modal. This has meant overriding the z-index in the navigation bar for
smaller screens.