Update controller and add views for Projects
A controller and resources routes for this resource had already been set
up but with no actions configured. Have updated the controller so that
it can be used by the front end and added HAML templates that can be
used for admin purposes as I still can't be bothered to add this to
the React front end.