Add ranks and published status to portfolio views
Added the ranks to Clients and Projects show pages and forms. Have also
updated the index so that only published projects are displayed. Have
also added a list view to both resources for admin use which will
display drafts as well as published projects.
This involved switching from ERB templates to cleaner HAML templates in
the clients resource.