Fix broken PostTag tests
Associations don't seem to work the way they used to in FactoryBot. Have
written a work around in the factories that should resolve this for now,
ensuring that the attributes defined in the `post` and `tag` factories
will be used for the associations in the `post_tag` factory. Have also
altered the validation in the PostTag model so that the uniqueness of
both foreign keys can be checked before a record is saved. Have made a
similar change to an attribute in the `comment` factory.