Remove rails_12factor gem
Removing this gem as this is no longer a requirement for deploying to
Heroku. This was phased out in RoR5 but now that I have upgraded to
RoR6, I am now concerned that it has caused caching issues with the
asset pipeline as it looks as though stylesheets aren't updating
anymore.