Fix Tag relationship methods
Have fixed the class methods in the Tag class so as to resolve a bug
that appears when an existing blog post is updated (this started to
appear after the RoR 6 upgrade). This should also greatly simplify the
way in which these methods behaved. In particular, I've removed a method
that checked for duplicates in the post_tags table as it's already
possible to validate for uniqueness on the the two coloumns in the
PostTag class.
Have added tests for the relevant methods as the test coverage wasn't
great for this class before this commit.