Make new tags case-insensitive
I want to be able to add tags that are capitalised where necessary
however I don't want a new tag to be created if, for example, a lower
case version of the tag is specified in the `tag_list` attribute. Have
added a new class method to perform a case-insensitive
find_or_create_by. If I have to add more ActiveRecord methods like this,
it may make sense to write a more generic module instead.