Add rich text body to the posts resource
Have created a migration to add a new `rich_text_body` colum to the
`posts` table and added the ActionText table the Posts form. The
posts#show page now displays ActionText/ActionContent instead of the
Coderayify markdown from 2017. This involved adding a helper method that
will truncate the post for the previews in the post#index page.