Fix Turbolinks in rails application
Removing jquery UJS from the javascript broke the delete links in any
admin views. Re-adding in order to fix bug.
Have also removed a javascript_packs tag from one of the posts view as
this is no longer needed (the one in the _head partial should add this
script as needed instead)