Attempting to fix Job specs
Have updated the factory so that the description uses an ActionText
type and made changes to the PUT test to accommodate this. These tests
still don't pass locally however.
Have also tweaked the CircleCI config slightly as it's suddenly unable
to see the ActionText tables and removed the schema file from gitignore
(even though it was somehow still being commited into git anyways).