Fix truncation on post#index
Switched to using the Nokogiri::HTML class (as opposed to the
Nokogiri::XML class) as this will return fragments of a tag instead
stopping at the last closed tag. I expect that this should improve click
rates (but idk).
Have also fixed the styling for code blocks in this view.