Add formatting for blog post dates
Have added new methods to handle the formatting of the published_at and
updated_at dates so that this logic can be moved out of the post views
and into the model. Have added tests and tweaked the styling. Publishing
dates should now appear in posts#index (which I think is a little better
for SEO but again idk).