Replace Trix editor with QuillJS in Posts
Am in the process of replacing the Trix editor in the /posts form as it
doesn't have most of the features I need and, to be quite honest,
ActionText feels a lot like overkill for this problem. QuillJS so far has
been buggy in that the icons don't render correctly locally. I am able to
save rich text with pre code blocks and images locally if I guess where
the icons are.