Add scope to /jobs resource
For some reason the devise helpers misbehave when applied to endpoints
called "/jobs". Essentially, in the local environment, the current
user's session was never visible in this controller and so it would
render a 404 page even when an admin was logged in. Renameing the
endpoints "/previous/jobs" seems to resolve the issue.