Hack to fix missing quill icons
The icons for the Quill toolbar have to be overwritten manually as,
prior to this change, all that was rendered was the path for each of the
SVG files. Am currently using XMLHttpRequest to read the files based on the
paths as attempting to use the js.erb files to read them caused more
problems than it solved (https://github.com/usabilityhub/rails-erb-loader/issues/63)
and the "fs" package just doesn't work in Rails/Webpacker (or if it does,
I wasn't using it right). This method appears to have been tagged for
deprecation however so I will need to find a better solution at some point.