Move javascript webpacker file back to the head tag
Moving this webpacker script to the head tag so in order to reduce
duplication across the editable forms. Will use the `quill_editor`
variable in the form partials to determine whether the page should call
the javascript file.
Have also tweaked the styling in the editor somewhat so that only one
scroll appears in the text area.