Add font sizes to text editor
Currently I have a number of header tags but obviously using these in
place of styled text is bad for SEO. Have added the font size button to
handle non-header text sizes. In the stylesheets, I've had to move the font
sizes to a mixin so that it can be used in both blog posts and comments.