Refactor User profiles
This commit hides the email addresses of users other than the current
user. This also fixes the redirect to the login page so that anonymous
users are redirected to a login page when they attempt to visit a user
profile rather than shown a 500 error. Have also fixed the formatting
for truncated comments that appeared on the page.
Have added specs accordingly.