Fix timezone issue
Locally a post_spec tests had failed as a result of my moving timezone.
I've updated the test so that it will take into account the timezone of
a given post and have added the server's timezone to timestamps in the
layout. In the future, I think that I'll need to put more thought into
how I handle timezones on the site but this will have to do for now.