Re-add missing javascript file to public
My error pages depend on a javascript file[1] that appears to have gone
missing at some point. This script would allow users to return to a
previous page if they received a 404, 422 or 500 error on the site. This
seems to have been missing for a while but has only recently started to
produce routing errors in the local environment today. This commit
should fix that problem.

[1]: https://github.com/kohrVid/kohrVid-website/blob/c678da9240870d400f11b55c1e51db09c7824670/public/javascripts/takeMeBack.js